Handle snake going OUT of the map
Updated Board to have methods to check if a position is within the map.

Updated Position to have + operators, so I can add a position and a direction
together.

Direction is now an Enum.

Updated StandardEngine to return a VALID random new position for the
snake.
Updated StandardEngineTests as well, to handle the new random positions
instead of a static "DOWN"