Handle logging within engines
Added logging to StandardEngine, which resulted in a lots of problems with
newing the class etc, from the factory.
The EngineFactory is no longer static, but is initialized with the
available engines, to make logging possible.
Updated GameController.cs and GameControllerTests.cs to reflect engine changes.
Also updated StandardEngineTests.cs to handle the logger.